fix_4329 Modify nodes and containerClient for the performance test
